+++ runner/webserver.ts
@@ -13,10 +13,11 @@
  */
 
 import * as cleankill from 'cleankill';
+import * as express from 'express';
 import * as fs from 'fs';
 import * as _ from 'lodash';
 import * as path from 'path';
-import {MainlineServer, PolyserveServer, RequestHandler, startServers, VariantServer} from 'polyserve';
+import {MainlineServer, PolyserveServer, RequestHandler, ServerOptions, startServers, VariantServer} from 'polyserve';
 import * as semver from 'semver';
 import * as send from 'send';
 import * as serverDestroy from 'server-destroy';
@@ -114,8 +115,7 @@ export function webserver(wct: Context): void {
       if (!version) {
         throw new Error(`
 The web-component-tester Bower package is not installed as a dependency of this project (${
-                                                                                           packageName
-                                                                                         }).
+            packageName}).
 
 Please run this command to install:
     bower install --save-dev web-component-tester
@@ -167,15 +167,22 @@ Expected to find a ${mdFilenames.join(' or ')} at: ${pathToLocalWct}/
     });
 
     // Serve up project & dependencies via polyserve
-    const polyserveResult = await startServers({
-      root: options.root,
-      compile: options.compile,
-      hostname: options.webserver.hostname,
-      headers: DEFAULT_HEADERS,
-      packageName,
-      additionalRoutes,
-      npm: !!options.npm,
-    });
+    const polyserveResult = await startServers(
+        {
+          root: options.root,
+          compile: options.compile,
+          hostname: options.webserver.hostname,
+          headers: DEFAULT_HEADERS,
+          packageName,
+          additionalRoutes,
+          npm: !!options.npm,
+        },
+        async (app: express.Express, options: ServerOptions) => {
+          await wct.emitHook(
+              'define:webserver', app, options,
+              async (substitution: express.Express) => app = substitution);
+          return app;
+        });
     let servers: Array<MainlineServer|VariantServer>;
 
     const onDestroyHandlers: Array<() => Promise<void>> = [];