+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 <!-- Add new, unreleased items here. -->
+
+## 6.3.0 - 2017-10-02
+
 * Updated wct-browser-legacy to use a module version of a11ySuite to get access to Polymer.dom.flush.
 * Updated generated index for webserver to use a11ySuite as a module.
+* Updated polyserve to get support for on-the-fly module compilation and `<script type=module>` conversion for browsers that don't support modules.
 
 ## 6.2.0 - 2017-09-19